Do not terminate the camera capturing when hiding.
This function is called e.g. when a scene is switched.
We do not want the camera to shut down at this time, since the next scene
may also include its image.

Furthermore, this fixes issue #12, though the underlying cause for the crash
still remains, namely that the terminate-funciton is called multiple times
without checking that the camera is still valid.